# openapi.yaml
openapi: 3.0.3
info:
  title: Club Operations POS API
  version: 1.0.0
  description: |
    Server-authoritative REST API for managing club check-ins, room/locker assignments,
    cleaning workflows, checkout verification, and operational reporting.

    All endpoints are versioned under `/v1`.
    Request bodies are validated with Zod schemas.

    Authentication:
    - Staff endpoints require Bearer token (session token from PIN or WebAuthn login)
    - Admin endpoints require ADMIN role in addition to authentication
    - Customer-facing endpoints (agreement signing, checkout kiosk) do not require auth
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /v1/cleaning/batch:
    post:
      tags: [Cleaning]
      summary: Batch update room statuses (DIRTY→CLEANING or CLEANING→CLEAN)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scanned]
              properties:
                scanned:
                  type: array
                  items:
                    type: object
                    required: [roomId, fromStatus, toStatus]
                    properties:
                      roomId:
                        type: string
                        format: uuid
                      fromStatus:
                        type: string
                        enum: [DIRTY, CLEANING, CLEAN]
                      toStatus:
                        type: string
                        enum: [DIRTY, CLEANING, CLEAN]
                      override:
                        type: boolean
                        default: false
                      overrideReason:
                        type: string
      responses:
        '200':
          description: Batch update completed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque session token from PIN or WebAuthn login (stored/validated server-side; not a JWT)
...